Normalise from path to allow symlink to a directory from inside that dir
It is currently not possible to symlink to a directory from inside that
directory (e.g. symlink /app/myfolder/folder to /app/myfolder). This is because
we're using path.dirname(to) which will result in the same directory as the from
argument and the result of using path.relative() with two paths that are the same
is '' - an empty string.

By using path.normalize() you get the correct result: '.', the current directory